+++ pyobjc-framework-Cocoa/Modules/_AppKit_protocols.m
@@ -282,5 +282,7 @@ static void __attribute__((__used__)) use_protocols(void)
     Py_XDECREF(p);
     p = PyObjC_IdToPython(@protocol(NSSpringLoadingDestination));
     Py_XDECREF(p);
+    p = PyObjC_IdToPython(@protocol(NSTextCheckingClient));
+    Py_XDECREF(p);
 #endif
 }