+++ docs/changelog.rst
@@ -10,6 +10,9 @@ Version 6.2.1
 
   Reported by Thomas Buchberger
 
+* Minor tweaks to build and pass tests on macOS 10.14 with the latest Xcode
+  that can be installed on that version of macOS.
+
 Version 6.2
 -----------